Delete invalid SIP users from Libki when using SIP authentication.
When SIP authentication sends back a invalid user flag, attempt to
delete that user in Libki. It is possible that the user attempting
to log in via Libki previously existed as a SIP user, but is no
longer. In this case, we don't want that user to exist in Libki
either, so we should attempt to delete the username, in case it
does exists in the Libki database.