Enhancement - Add ability to disallow logins if patron has exceeded fee limits when using SIP2
If a Libki server is connecting to a SIP2 server for authentication, this
feature allows the library to define a new parameter fee_limit in the SIP section
of the Libki configuration file. The value of this parameter may be either a fee limit
( e.g. 5.00 ) or a SIP2 field that should contain the fee limit as sent from the server
( e.g. CC ).