Add the ability to give users multiple sessions daily
This patch adds a daily allotment of minutes for each user, set in the existing
setting DefaultTimeAllowance.

A new setting DefaultSessionTimeAllowance is added, defaulting to DefaultTimeAllowance
to keep existing behavior.

By default each night a user is given a daily allotment of minutes. These minutes are
then transferred into the session minutes up to the maximum session time allowed.

When a user is logged in, his or her session minutes decrements until it reaches 0
at which time the user is logged out. The user may also voluntarily log out early.

Thus, if the allowance per day is 60, and the allowance per session is 15, a given patron
will be able to log into a Libki client 4 times per day, for 15 minutes at a time.

Any logged out user who has daily minutes remaining and less than the full amount
of session minutes will have minutes transferred from the daily allotment to the
session allotment.

For example, using the above settings, a user starts out with 15 minutes in his or her
session time, and 45 minutes in his or her daily allotment ( 15 + 45 = 60 ).
This user then logs in for a full 15 minutes, and reaches 0 session minutes. He or she
is then logged out, and at that time, 15 mintues is transferred from the daily allotment
to the session allotment and the user now has 30 minutes left in the daily allotment and 15
for the next session. The user now logs in again for 10 minutes and then logs out. At
this point the system will transfer 10 minutes more from the daily pool to the session pool
such that the patron has 20 minutes left in the daily allotment and 15 minutes ready for
the next session.