Add i18n to Libki Server.
Sponsored by: Ville de Victoriaville

Internationalization done with Catalyst::Plugin::I18N

This feature includes:

All string literals have been externalized in po and pot files
Strings that appear in javascript code have to be carefully written, as you need to escape symbols properly.
(Ex: ' would end the string in the js code, the .tt is interpreted server-side)

Added controller I18N.pm.
The new sub "changelocale" is used to store a new locale in the session.

Room.pm's "end" sub takes the locale stored in the session and feeds it in i18n's "languages" sub.

Added a new inlcude "language_menu" which is a collapsable bootstrap menu with links to "changelocale" that are generated using i18n's "installed_languages" sub.

The new menu has been added to both navbars.

Localization for the datatables has been done with the plugin's own language files.
The file is fetched directly from their server using ajax when initializing the table.
The code can, in theory, fetch the correct language file as is uses the language string stored by the i18n Catalyst plugin. (It luckily matches the syntax used in datatable's urls!)

Since the breadcrumb is dynamically generated using the url, I have changed the label generation subroutine to pass the label to the i18n localization sub.
The keys follow this pattern "breadcrumb.«page»".
They need to be manualy added to the po files when a new page is added to the website.

Same goes for the languages shown under the language menu.
The keys are: "lang.«language_code»" and need to be added in the default po when adding a new language.
This way the menu items' text are displayed in the correct language, not only english.
French is Français, English is... English and Spanish would be Español (For example)

The .tt [% META ... %] instructions have been changed to [% meta.title = c.loc(...   %], because the former tag cannot contain code, only literals, and we want to translate the title.

I can't seem to get Unicode to work correctly... so all the accents are represented in ugly html and hexa codes. In the french fr.po file, of course.

	modified:   lib/Catalyst/Plugin/Breadcrumbs.pm
	modified:   lib/Libki.pm
	new file:   lib/Libki/Controller/I18N.pm
	modified:   lib/Libki/Controller/Root.pm
	new file:   lib/Libki/I18N/fr.po
	new file:   lib/Libki/I18N/i_default.po
	new file:   lib/Libki/I18N/messages.pot
	new file:   root/dynamic/includes/language_menu.tt
	modified:   root/dynamic/includes/navbar_administration.tt
	modified:   root/dynamic/includes/navbar_public.tt
	modified:   root/dynamic/includes/wrapper.tt
	modified:   root/dynamic/templates/administration/history/index.tt
	modified:   root/dynamic/templates/administration/history/statistics.tt
	modified:   root/dynamic/templates/administration/index.tt
	modified:   root/dynamic/templates/administration/settings/index.tt
	modified:   root/dynamic/templates/login.tt
    modified:   root/dynamic/templates/public/index.tt