Switch admin settings to using English strings for translation.
The idea of using labels for each and every string in Libki is just not working.
Instead, we should be putting the English version in the c.loc calls and allow
that to be translated directly. The should improve both ease of development and
ease of translation.

I will be working on switching over all the templates to this system at which
time I will make a request for updated translations!