Add ability to specify deny_on_field value
This is an enhancement to the SIP2 configuration option deny_on_field.
With this change, one can now specify not just a field and reason for denial,
but a value to deny on as well. Previously, deny_on_field assumed that a login
should be denied if a field had a value of 'Y'. Now, the admin can specify
what value a field should have in order to trigger a denial.

So, the following will continue to default to the deny on value of 'Y':
deny_on_field AB:This is the reason we are denying you
while this will assume a login should only be denied of the the value
of the field is 'ABC':
deny_on_field AF:This is the reason we are denying you:ABC