Issue 38: Add auto deletion of unused users
Over in Europe everything is about GDPR (a new personal data protection legislation) these days. One of the things that's pretty much demanded is some form of deletion of user data that isn't being used. For us, that'd be accounts. I realised this during the weekend, so the chance of it being in place for the new law (May 25) is kind of limited, but this will need to be fixed.

Here's my idea:

Add a last_registred timestamp field in the users table.
Update that value whenever a user uses a client.
A GUI option in Settings that sets the delete users-time in amount of X days.
Deleting the users that are to be dropped (i. e. haven't been active within X days) with the nightly script.

https://github.com/Libki/libki-server/issues/38