Add ability to skip password verification when using SIP2
This is useful if you don't want to use passwords, or if your
SIP server doesn't send BL fields so you cannot ascertain the
veracity of the given password.

It should be noted that the client has an INI option to hide
the password field from the login screen. This feature and that
one could work well in conjunction with one another.