Move session minutes from users table to sessions table
This is a big change for Libki. In the beginning, Libki only had one minutes column.
Session minutes and daily minutes were one and the same ( with no automatic way to gain more minutes ).
Then we added the daily allotment to refill the session minutes whenever a patron was logged out.
Now we've come to the point where the need for flexibility in defining sesson ( and daily ) minutes
means that having a pre-defined number of session minutes is a limit.

Let's say we want to limit the number of session minutes by the workstation location.
There is no really good way to do that right now, because when you refill session minutes
from daily minutes, you care not logged in. No login means no session means no workstation
means no workstation location.

Instead, we will now set the session minutes at the moment of successful login, and move
the session minutes column from the users table to the sessions table!

This will allow us to have more dynamic rules for determining session minutes in future developments!