Add ability to specify session length by user category and/or client location
This adds an 'advanced' mode where a set of rules can be defined as YAML.

For example:
-
  criteria:
    user_category: MyCategory1
    client_location: MyLocation1
  rules:
    session: 25
    guest_session: 20
-
  criteria:
    client_location: MyLocation1
  rules:
    session: 20
    guest_session: 20
-
  criteria:
    user_category: MyCategory1
  rules:
    session: 15
    guest_session: 20
-
  rules:
    session: 10
    guest_session: 20

Rules should be list from most specific to least specific.
If a criteria is missing in a rule, then it is not used as part if the matching process.

For example, let's say we are attempting to log in as a patron with the category MyCategory1, and a client with the location MyLocation2.
The first rule matches category, but not location, so it is skipped.
The second rule only checks location, which is not a match, so it is skipped.
The third rule only checks category, which matches, so it is used!
The last rule has no critea, so it matches on *everything*

If there is no catch-all rule like this, Libki will fall back to the origin default time allowance settings

NOTE: Even if a rule should match, it will be skipped if it doesn't contain the 'subrule' we are looking for.
For example, take the following:
-
  criteria:
    client_location: MyLocation1
  rules:
    session: 15
-
  rules:
    session: 10
    guest_session: 20

Assume a guest is logging in to a MyLocation1 client.
The first rule will be skipped because guest_session is not defined in this rule.
The second rule will be used because it matches all and *does* have guest_session defined.