Add ability to specify daily minutes with advanced rules
This is another major change regaurding minutes. To be able to set
daily minutes by location, we need to wait until a patron logs in for
the first time that day.

It makes no functional difference if daily minutes are defined each night
or at first login time. Either way they are set for the day.

At this point, the minutes field in the patron editor defines sets
the daily allotment, where the time editor on the clients table
sets session minutes. This makes good sense imho.