Allow rules to specify multiple values for a given rule criteria
For example:
-
  criteria:
    user_category:
      - MyCategory1
      - MyCategory2
    client_location: MyLocationA
  rules:
    session: 25
    daily: 50
    guest_session: 20
    guest_daily: 45

would apply to any user of category MyCategory1 *or* MyCategory2 logging into a client at locatin MyLocationA.

This is, this rule will apply to all the value in the list for that criteria.

Another way to think of it is that all creteria are ANDed, and all the value for a criteria are ORed.

-
  criteria:
    user_category:
      - Cat1
      - Cat2
    client_location:
      - LocA
      - LocB
  rules:
    session: 25
    daily: 50
    guest_session: 20
    guest_daily: 45

Would be express like:
SELECT RULE WHERE ( user_category = Cat1 OR user_category = Cat2 ) AND ( client_location = LocA OR client_location = LocB )