New utility tool added - Translation updater
It has two functions.

1. It can generate a fresh messages.pot file (that's the one all the translations are based on). This is done by passing --generate or -g for short.

2. It can update a translation file to the current messages.pot file. This is done by passing --update lib/Libki/I18N/filename or -u lib/Libki/I18N/filename for short.

Since this is a dev tool I didn't put all the bells and whistles into it - i e it doesn't check if the file passed actually is a translation file and it's a bit bulky since it's working on files in lib/Libki/I18N but has to be run from libki-server.

I still think this is all right since it's not something you run every other day and it's a whole lot better than doing it all by hand.