Disable CI for Pull Requests
Pull requests always fail CI because of how the scripts work. For now we shoulud just disable CI on pull requests.

In the future we should move as much logic as we can for the build script into actual shell scripts.
Then we can run the image build and unit tests all at once for pull requests without pushing an artifact.
For the master branch, we can continue to keep separate stages where we build the docker image, push it, then test it.