Fix error preventing addding admin priveleges to new users
If a new user is created and one then tries to set admin permissions
for that user, it fails with the server generating the message:
ERROR - Caught exception in
Libki::Controller::Administration::API::User->update "Can't call method
"delete" on an undefined value at
/app/script/../lib/Libki/Controller/Administration/API/User.pm line
300."

This is because we are using find, but the user has no minutes allotment
yet. Switching from find() to search() resolves the issue.