Issue #376:  Add date range filter to History page
History can get pretty lengthy, and while you can search for dates in
the search bar, it's keyword not date range.  This commit adds a filters
columns to the History page, and puts date range (from/to) filtering
in as the first filter available.  More can be added based on need