Issue #383:  Norwegian and Portuguese translations not working in DataTables
Basically, languages are complicated, and using the more specific form of
nb_NO and pt_PT for the names of our PO files meant that I18N::LangTags
couldn't find the common name for the language, which broke fetching from
the DataTables CDN.

Moved both .po files

Also reworked the languages menu translations.  The word used should be the
one the language uses for itself, as this is the entry point into using that
language, and presuming that a user is familiar with the name of their language
in another is not terribly fair.  Keeps this in the .po files, though, so it
could be overridden locally more easily