Issue #407:  UTF-8 encoding issues
Up to this point, we were using Catalyst Runtime 5.90011, which is
VERY old.  After 5.90080, UTF-8 character handling was baked in, so
first off, this patch upgrades to the latest stable runtime, 5.90132

This removes the need to manually decode fields before presenting them
to the View.  The bulk of this commit is removing calls to Encode::decode()
throughout the Controllers.

We also need to be sure that our DB connection is UTF-8.  This was added to
the libki_local.conf example file, but not the Docker version, so that's been
updated for completeness.

Depending on your database server configurations, you may need to adjust your
table character sets and collate to utf8mb4.  Exercise left to reader