Issue #384:  Add view option to public printing interface
Adds a 'View' button along with Print and Cancel buttons on the
list of print jobs for a user.  Opens in a new browser window.
Access is limited to the user who owns the print job

This is especially helpful now that we can do automated PDF conversion;
users may want to verify that their file rendered right before paying
to print it.