Issue #386: save changes to print job values instantly
On the public interface, a user can now change their printer independently
of issuing the print command.  This commit also enables the ability to change
the number of copies.

Creates a new API endpoint /public/api/user/update_print_job, which accepts
GET parameters for number of copies and/or printer.  Copies must be numeric
and greater than 0, and printer names must match ones that exist in the
configuration