Sync lib and installer directories back when running docker-compose
Scripts like development/update_schema.sh and utilities/translate.sh
make changes to code in lib and installer as part of their process.
When run in a container, this data is lost upon restart unless a
volume is mounted.  This commit does just that, allowing developers
to run these scripts with docker exec, and get the outputs back to
the source Git repository