Issue #418: Implement Transactions table
Creates a new entry in the Libki Schema for Transactions, which
store financial transactions (credits/debits) for Users.  This supports
the addition of electronic self-pay solutions, but also provides a more
specialized structure for what Issue #365 initially did.

Migrates the funds value currently in users.funds to a transaction line.

Keep in mind that transactions are stored as cents rather than dollars
(with multi-currency support enabled for later); this helps to prevent
floating point errors.  Conversion to dollars will be done with legacy
API endpoints to ensure backwards compatibility.