use 'rebuild' instead of 'sync+restart' for docker-compose watch
Re-enables the volume mounts on lib and installer so that development
scripts work.  Changes the action on watch to rebuild, which now that
the order of operations in the Dockerfile is changed, won't cause
a full CPAN module reinstall every time.  Only takes a little longer
than sync+restart.

Ignores docker and ansible directories, as those seem a bit too
meta to be including in this process